Update: Modify catalog api

+++ src/main/java/com/yas/system/catalog/internal/service/BrandService.java
@@ -8,9 +8,9 @@ import com.yas.system.common.response.PageResponse;
 
 public interface BrandService {
     void createBrand(BrandCreateRequest request);
-    void updateBrand(BrandUpdateRequest request, Long brandId);
-    void deleteBrandById(Long brandId);
-    BrandResponse getById(Long brandId);
+    void updateBrand(BrandUpdateRequest request, Integer brandId);
+    void deleteBrandById(Integer brandId);
+    BrandResponse getById(Integer brandId);
     PageResponse<BrandResponse> getBrandPage(Integer pageNumber, Integer pageSize);
 }
